msr: add support for ATOM_SILVERMONT_D (model 0x4D)
The DComp Testbed has a Silicom RCC-VE network board with an Intel Atom
C2558 processor. It seems we can read RAPL settings, but not set them on
this particular platform, even though the zones are not locked (PACKAGE
and CORE MSR fields simply won't update).
This CPU line is fairly old and for some reason not actually supported by
the Linux kernel's Intel RAPL powercap driver. Maybe the inability to set
constraint values is related.